refactor: [#116] spec 주소 오류 수정
