+++ back/src/post-contents/entities/post-content.entity.ts
@@ -1,6 +1,12 @@
 import { Content } from 'src/contents/entities/content.entity';
 import { Post } from 'src/post/entities/post.entity';
-import { Entity, JoinColumn, ManyToOne, OneToOne, PrimaryColumn } from 'typeorm';
+import {
+  Entity,
+  JoinColumn,
+  ManyToOne,
+  OneToOne,
+  PrimaryColumn,
+} from 'typeorm';
 
 @Entity({ name: 'post_contents' })
 export class PostContent {
