+++ back/src/post/spec/post.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { PostController } from './post.controller';
-import { PostService } from './post.service';
+import { PostController } from '../post.controller';
+import { PostService } from '../post.service';
 
 describe('PostController', () => {
   let controller: PostController;
