+++ back/src/post/spec/post.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { PostService } from './post.service';
+import { PostService } from '../post.service';
 
 describe('PostService', () => {
   let service: PostService;
